Modify formatting of form section
Changed several font settings of the form header "Let's do this!"
including color, a resizing and spacing. The font was way too close to
the rest of the form prior. Also creating more spacing between each
input section of the form which also partially solved the issue of the
form section itself being too small on the page.