Modify formatting of inputs in form section
Moved the labels for each part of the form on top of their respective
inputs per the project instructions. Form section now takes up 80% of
the entire page and even stretches it beyond the viewport a bit,
leaving a scrollbar in it's wake. Have an idea to fix this in the HTML
itself in a future update.